+++ docs/ReleaseNotes.md
@@ -4,6 +4,10 @@
 
 .netstandard2.0 based also support for netfx4.5 common helpers and extensions
 
+### [WeihanLi.Common 1.0.29](https://www.nuget.org/packages/WeihanLi.Common/1.0.29)
+
+- add di extension
+- add `StringExtension.TrimStart(this string str, string start)` extension
 
 ### [WeihanLi.Common 1.0.28](https://www.nuget.org/packages/WeihanLi.Common/1.0.28)